+++ package.json
@@ -1,0 +1,56 @@
+{
+  "name": "coder-back",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {
+    "abbrev": "^1.1.1",
+    "anymatch": "^3.1.2",
+    "balanced-match": "^1.0.2",
+    "binary-extensions": "^2.2.0",
+    "brace-expansion": "^1.1.11",
+    "braces": "^3.0.2",
+    "chokidar": "^3.5.3",
+    "concat-map": "^0.0.1",
+    "debug": "^3.2.7",
+    "dotenv": "^16.0.3",
+    "express": "^4.18.2",
+    "fill-range": "^7.0.1",
+    "glob-parent": "^5.1.2",
+    "has-flag": "^3.0.0",
+    "ignore-by-default": "^1.0.1",
+    "is-binary-path": "^2.1.0",
+    "is-extglob": "^2.1.1",
+    "is-glob": "^4.0.3",
+    "is-number": "^7.0.0",
+    "minimatch": "^3.1.2",
+    "ms": "^2.1.3",
+    "nodemon": "^2.0.20",
+    "nopt": "^1.0.10",
+    "normalize-path": "^3.0.0",
+    "picomatch": "^2.3.1",
+    "pstree.remy": "^1.1.8",
+    "readdirp": "^3.6.0",
+    "semver": "^5.7.1",
+    "simple-update-notifier": "^1.0.7",
+    "supports-color": "^5.5.0",
+    "to-regex-range": "^5.0.1",
+    "touch": "^3.1.0",
+    "undefsafe": "^2.0.5"
+  },
+  "scripts": {
+    "start": "node index.js",
+    "dev": "nodemon index.js",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ignacioborraz/coder-back.git"
+  },
+  "author": "igna",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/ignacioborraz/coder-back/issues"
+  },
+  "homepage": "https://github.com/ignacioborraz/coder-back#readme"
+}
